refactor(cr): create domain-based file-structure
