+++ src/app/collective-rewards/builders/context/BuilderContext.tsx
@@ -2,7 +2,7 @@ import { createContext, FC, ReactNode, useContext, useMemo } from 'react'
 import { Address } from 'viem'
 import { BuilderStatus } from '@/app/collective-rewards/types'
 import { useGetProposalsState } from '@/app/collective-rewards/whitelist/hooks/useGetProposalsState'
-import { useGetBuilders } from '@/app/collective-rewards/hooks'
+import { useGetBuilders } from '@/app/collective-rewards/builders'
 import { getMostAdvancedProposal } from '@/app/collective-rewards/utils/getMostAdvancedProposal'
 import { DateTime } from 'luxon'
 import { splitCombinedName } from '@/app/proposals/shared/utils'
