+++ src/app/collective-rewards/builders/hooks/index.ts
@@ -1,6 +1,4 @@
 export * from './useGetBuilders'
 export * from './useGetIsWhitelistedBuilder'
-export * from './useGetRewardDistributedLogs'
-export * from './useGetTokenProjectedReward'
 export * from './useGetWhitelistedBuilders'
 export * from './useGetWhitelistedBuildersLength'
