+++ src/app/collective-rewards/builders/hooks/useGetBuilders.ts
@@ -1,4 +1,7 @@
-import { useGetIsWhitelistedBuilder, useGetWhitelistedBuilders } from '@/app/collective-rewards/hooks'
+import {
+  useGetIsWhitelistedBuilder,
+  useGetWhitelistedBuilders,
+} from '@/app/collective-rewards/builders/hooks'
 import { BuilderInfo } from '@/app/collective-rewards/types'
 import { useFetchCreateBuilderProposals } from '@/app/proposals/hooks/useFetchLatestProposals'
 import { useMemo } from 'react'
