+++ src/app/collective-rewards/builders/hooks/useGetIsWhitelistedBuilder.ts
