+++ src/app/collective-rewards/leaderboard/hooks/index.ts
@@ -1,0 +1,1 @@
+export * from './useGetBuildersRewards'
