+++ src/app/collective-rewards/leaderboard/index.ts
@@ -1,0 +1,2 @@
+export * from './hooks'
+export * from './LeaderBoard'
