+++ src/app/collective-rewards/metrics/components/CycleMetrics.tsx
