+++ src/app/collective-rewards/rewards/hooks/useClaimAllRewards.ts
