+++ src/app/collective-rewards/rewards/index.ts
@@ -1,0 +1,2 @@
+export * from './hooks'
+export * from './MyRewards'
