+++ src/app/collective-rewards/whitelist/components/Grid/index.ts
@@ -1,0 +1,2 @@
+export * from './WhitelistGrid'
+export * from './WhitelistGridItem'
