+++ src/app/collective-rewards/whitelist/components/Search/index.ts
@@ -1,0 +1,1 @@
+export * from './WhitelistSearch'
