+++ src/app/collective-rewards/whitelist/components/index.ts
@@ -1,0 +1,2 @@
+export * from './Grid'
+export * from './Search'
