+++ src/app/collective-rewards/whitelist/context/WhitelistContext.tsx
@@ -1,7 +1,7 @@
+import { BuilderProposal, withBuilderContextProvider } from '@/app/collective-rewards/builders'
 import { BuilderStatus } from '@/app/collective-rewards/types'
+import { useGetFilteredBuilders } from '@/app/collective-rewards/whitelist'
 import { createContext, Dispatch, FC, ReactNode, SetStateAction, useContext, useState } from 'react'
-import { useGetFilteredBuilders } from '@/app/collective-rewards/whitelist/hooks/useGetFilteredBuilders'
-import { BuilderProposal, withBuilderContextProvider } from '@/app/collective-rewards/BuilderContext'
 
 export type BuilderStatusFilter = 'all' | BuilderStatus
